FFX: Fix dynamic draws from monsters.
To try to simplify rendering, I had everything handled by the current
scene (relevantly, the subscene during battle). However, particle
renderers allocated dynamic draws (e.g. electricity) based on which
scene created the underlying data, which was always the main scene
(because I wanted monster data to be able to be reused).
So now the dynamic buffer managers have to live on the particle system,
rather than with the data.

This meant threading yet another piece of graphics state through the
particle rendering functions, so now those are stuffed in a new
temporary context object. This will create a few pieces of garbage per
frame, which hopefully isn't a big deal.